Refactor - split up functions into files
