+++ setup.cfg
@@ -1,0 +1,7 @@
+[flake8]
+max-line-length = 120
+exclude =
+    .git,
+    __pycache__,
+    build,
+    dist
